Reference

Reference

Index

Autologistic.ALRsimple

Autologistic.ALRsimple

Autologistic.ALRsimple

Autologistic.ALfit

Autologistic.ALsimple

Autologistic.ALsimple

Autologistic.ALsimple

Autologistic.ALsimple

Autologistic.AbstractAutologisticModel

Autologistic.AbstractPairwiseParameter

Autologistic.AbstractUnaryParameter

Autologistic.CenteringKinds

Autologistic.FullPairwise

Autologistic.SamplingMethods

Autologistic.fullPMF

Autologistic.sample

Types and Constructors

Autologistic.ALRsimple — Type.

ALRsimple

An autologistic regression model with "simple smoothing": the unary parameter is of type LinPredUnary, and the pairwise parameter is of type SimplePairwise.

source

Autologistic.ALRsimple — Method.

ALRsimple(
    unary::LinPredUnary,
    pairwise::SimplePairwise;
    Y::Union{Nothing,<:VecOrMat}=nothing,
    centering::CenteringKinds=none, 
    coding::Tuple{Real,Real}=(-1,1),
    labels::Tuple{String,String}=("low","high"), 
    coordinates::SpatialCoordinates=[(0.0,0.0) for i=1:size(unary,1)]
)

Constructs an ALRsimple autologistic regression model, with unary part unary and pairwise part pairwise.

Keyword arguments

Y is the array of dichotomous responses. Any array with 2 unique values will work. If the array has only one unique value, it must equal one of th coding values. The supplied object will be internally represented as a Boolean array.

centering controls what form of centering to use.

coding determines the numeric coding of the dichotomous responses.

labels is a 2-tuple of text labels describing the meaning of Y. The first element is the label corresponding to the lower coding value.

coordinates is an array of 2- or 3-tuples giving spatial coordinates of each vertex in the graph. Default is to set all coordinates to zero.

Examples

julia> u = LinPredUnary(rand(10,3));
julia> p = SimplePairwise(Graph(10,20));
julia> model = ALRsimple(u, p, Y = rand([-2, 3], 10))
Autologistic regression model of type ALRsimple with parameter vector [β; λ].
Fields:
  responses    10×1 Bool array
  unary        10×1 LinPredUnary with fields:
                 X  10×3×1 array (covariates)
                 β  3-element vector (regression coefficients)
  pairwise     10×10×1 SimplePairwise with fields:
                 λ      [0.0] (association parameter)
                 G      the graph (10 vertices, 20 edges)
                 count  1 (the number of observations)
                 A      10×10 SparseMatrixCSC (the adjacency matrix)
  centering    none
  coding       (-1, 1)
  labels       ("low", "high")
  coordinates  10-element vector of Tuple{Float64,Float64}

source

Autologistic.ALRsimple — Method.

ALRsimple(
    graph::SimpleGraph{Int}, 
    X::Float2D3D; 
    Y::VecOrMat=Array{Bool,2}(undef,nv(graph),size(X,3)),
    β::Vector{Float64}=zeros(size(X,2)),
    λ::Float64=0.0, 
    centering::CenteringKinds=none, 
    coding::Tuple{Real,Real}=(-1,1),
    labels::Tuple{String,String}=("low","high"),
    coordinates::SpatialCoordinates=[(0.0,0.0) for i=1:nv(graph)]
)

Constructs an ALRsimple autologistic regression model from a graph and an X array of predictors. Use Lightgraphs.jl functions to make the graph. X is n×p×m, where n is the number of vertices in the graph, p is the number of predictors, and m is the number of (multivariate) observations. If m == 1, X can be 2D.

Keyword arguments

Y is the array of dichotomous responses. Any array with 2 unique values will work. If the array has only one unique value, it must equal one of th coding values. The supplied object will be internally represented as a Boolean array.

β is the regression coefficients.

λ is the association parameter.

centering controls what form of centering to use.

coding determines the numeric coding of the dichotomous responses.

labels is a 2-tuple of text labels describing the meaning of Y. The first element is the label corresponding to the lower coding value.

coordinates is an array of 2- or 3-tuples giving spatial coordinates of each vertex in the graph. Default is to set all coordinates to zero.

Examples

julia> model = ALRsimple(Graph(10,20), rand(10,3), β=[0.5, 0.75, 1.0], λ=1.1)
Autologistic regression model of type ALRsimple with parameter vector [β; λ].
    Fields:
      responses    10×1 Bool array
      unary        10×1 LinPredUnary with fields:
                     X  10×3×1 array (covariates)
                     β  3-element vector (regression coefficients)
      pairwise     10×10×1 SimplePairwise with fields:
                     λ      [1.1] (association parameter)
                     G      the graph (10 vertices, 20 edges)
                     count  1 (the number of observations)
                     A      10×10 SparseMatrixCSC (the adjacency matrix)
      centering    none
      coding       (-1, 1)
      labels       ("low", "high")
      coordinates  10-element vector of Tuple{Float64,Float64}

source

Autologistic.ALfit — Type.

ALfit

A type to hold estimation output for autologistic models. Fitting functions return an object of this type.

Depending on the fitting method, some fields might not be set. Fields that are not used are set to nothing or to zero-dimensional arrays. The fields are:

estimate: A vector of parameter estimates.

se: A vector of standard errors for the estimates.

pvalues: A vector of p-values for testing the null hypothesis that the parameters equal zero (one-at-a time hypothesis tests).

CIs: A vector of 95% confidence intervals for the parameters (a vector of 2-tuples).

optim: the output of the call to optimize used to get the estimates.

Hinv (used by fit_ml!): The inverse of the Hessian matrix of the objective function, evaluated at the estimate.

nboot (fit_pl!): number of bootstrap samples to use for error estimation.

kwargs (fit_pl!): holds extra keyword arguments passed in the call to the fitting function.

bootsamples (fit_pl!): the bootstrap samples.

bootestimates (fit_pl!): the bootstrap parameter estimates.

convergence: either a Boolean indicating optimization convergence ( for fit_ml!), or a vector of such values for the optimizations done to estimate bootstrap replicates.

The empty constructor ALfit() will initialize an object with all fields empty, so the needed fields can be filled afterwards.

source

Autologistic.ALsimple — Type.

ALsimple

An autologistic model with a FullUnary unary parameter type and a SimplePairwise pairwise parameter type. This model has the maximum number of unary parameters (one parameter per variable per observation), and a single association parameter.

source

Autologistic.ALsimple — Type.

ALsimple(graph::SimpleGraph{Int}, count::Int=1; ...)

Construct an ALsimple autologistic model from a graph, with count undefined observations. Keyword arguments are the same as the ALsimple(graph::SimpleGraph{Int}, alpha::Float1D2D; ...) constructor.

Examples

julia> g = Graph(10,20);
julia> model1 = ALsimple(g);
julia> model2 = ALsimple(g, 4);
julia> size(model1.responses)
(10, 1)

julia> size(model2.responses)
(10, 4)

source

Autologistic.ALsimple — Method.

ALsimple(
    unary::FullUnary, 
    pairwise::SimplePairwise; 
    Y::Union{Nothing,<:VecOrMat}=nothing, 
    centering::CenteringKinds=none, 
    coding::Tuple{Real,Real}=(-1,1),
    labels::Tuple{String,String}=("low","high"), 
    coordinates::SpatialCoordinates=[(0.0,0.0) for i=1:size(unary,1)]
)

Constructs an ALsimple autologistic model with unary part unary and pairwise part pairwise.

Keyword arguments

Y is the array of dichotomous responses. Any array with 2 unique values will work. If the array has only one unique value, it must equal one of th coding values. The supplied object will be internally represented as a Boolean array.

centering controls what form of centering to use.

coding determines the numeric coding of the dichotomous responses.

labels is a 2-tuple of text labels describing the meaning of Y. The first element is the label corresponding to the lower coding value.

coordinates is an array of 2- or 3-tuples giving spatial coordinates of each vertex in the graph. Default is to set all coordinates to zero.

Examples

julia> u = FullUnary(rand(10));
julia> p = SimplePairwise(Graph(10,20));
julia> model = ALsimple(u, p)
Autologistic model of type ALsimple with parameter vector [α; λ].
Fields:
  responses    10×1 Bool array
  unary        10×1 FullUnary with fields:
                 α  10×1 array (unary parameter values)
  pairwise     10×10×1 SimplePairwise with fields:
                 λ      [0.0] (association parameter)
                 G      the graph (10 vertices, 20 edges)
                 count  1 (the number of observations)
                 A      10×10 SparseMatrixCSC (the adjacency matrix)
  centering    none
  coding       (-1, 1)
  labels       ("low", "high")
  coordinates  10-element vector of Tuple{Float64,Float64}

source

Autologistic.ALsimple — Method.

ALsimple(
    graph::SimpleGraph{Int}, 
    alpha::Float1D2D; 
    Y::VecOrMat=Array{Bool,2}(undef,nv(graph),size(alpha,2)), 
    λ::Float64=0.0, 
    centering::CenteringKinds=none, 
    coding::Tuple{Real,Real}=(-1,1),
    labels::Tuple{String,String}=("low","high"),
    coordinates::SpatialCoordinates=[(0.0,0.0) for i=1:nv(graph)]
)

Constructs an ALsimple autologistic model from a graph and unary parameter values alpha.

Keyword arguments

Y is the array of dichotomous responses. Any array with 2 unique values will work. If the array has only one unique value, it must equal one of th coding values. The supplied object will be internally represented as a Boolean array.

λ is the association parameter.

centering controls what form of centering to use.

coding determines the numeric coding of the dichotomous responses.

labels is a 2-tuple of text labels describing the meaning of Y. The first element is the label corresponding to the lower coding value.

coordinates is an array of 2- or 3-tuples giving spatial coordinates of each vertex in the graph. Default is to set all coordinates to zero.

Examples

julia> model = ALsimple(Graph(10,20), rand(10))
Autologistic model of type ALsimple with parameter vector [α; λ].
Fields:
  responses    10×1 Bool array
  unary        10×1 FullUnary with fields:
                 α  10×1 array (unary parameter values)
  pairwise     10×10×1 SimplePairwise with fields:
                 λ      [0.0] (association parameter)
                 G      the graph (10 vertices, 20 edges)
                 count  1 (the number of observations)
                 A      10×10 SparseMatrixCSC (the adjacency matrix)
  centering    none
  coding       (-1, 1)
  labels       ("low", "high")
  coordinates  10-element vector of Tuple{Float64,Float64}

source

Autologistic.AbstractAutologisticModel — Type.

AbstractAutologisticModel

Abstract type representing autologistic models. All concrete subtypes should have the following fields:

responses::Array{Bool,2} – The binary observations. Rows are for nodes in the graph, and columns are for independent (vector) observations. It is a 2D array even if there is only one observation.

unary<:AbstractUnaryParameter – Specifies the unary part of the model.

pairwise<:AbstractPairwiseParameter – Specifies the pairwise part of the model (including the graph).

centering<:CenteringKinds – Specifies the form of centering used, if any.

coding::Tuple{T,T} where T<:Real – Gives the numeric coding of the responses.

labels::Tuple{String,String} – Provides names for the high and low states.

coordinates<:SpatialCoordinates – Provides 2D or 3D coordinates for each vertex in the graph.

This type has the following functions defined, considered part of the type's interface. They cover most operations one will want to perform. Concrete subtypes should not have to define custom overrides unless more specialized or efficient algorithms exist for the subtype.

getparameters and setparameters!

getunaryparameters and setunaryparameters!

getpairwiseparameters and setpairwiseparameters!

centeringterms

negpotential, pseudolikelihood, and loglikelihood

fullPMF, marginalprobabilities, and conditionalprobabilities

fit_pl! and fit_ml!

sample and oneboot

showfields

Examples

julia> M = ALsimple(Graph(4,4));
julia> typeof(M)
ALsimple{CenteringKinds,Int64,Nothing}
julia> isa(M, AbstractAutologisticModel)
true

source

Autologistic.AbstractPairwiseParameter — Type.

AbstractPairwiseParameter

Abstract type representing the pairwise part of an autologistic regression model.

All concrete subtypes should have the following fields:

G::SimpleGraph{Int} – The graph for the model.

count::Int – The number of observations.

In addition to getindex() and setindex!(), any concrete subtype P<:AbstractPairwiseParameter should also have the following methods defined:

getparameters(P), returning a Vector{Float64}

setparameters!(P, newpar::Vector{Float64}) for setting parameter values.

Note that indexing is performance-critical and should be implemented carefully in subtypes.

The intention is that each subtype should implement a different way of parameterizing the association matrix. The way parameters are stored and values computed is up to the subtypes.

This type inherits from AbstractArray{Float64, 3}. The third index is to allow for multiple observations. P[:,:,r] should return the association matrix of the rth observation in an appropriate subtype of AbstractMatrix. It is not intended that the third index will be used for range or vector indexing like P[:,:,1:5] (though this may work due to AbstractArray fallbacks).

Examples

julia> M = ALsimple(Graph(4,4));
julia> typeof(M.pairwise)
SimplePairwise
julia> isa(M.pairwise, AbstractPairwiseParameter)
true

source

Autologistic.AbstractUnaryParameter — Type.

AbstractUnaryParameter

Abstract type representing the unary part of an autologistic regression model.

This type inherits from AbstractArray{Float64, 2}. The first dimension is for vertices/variables in the graph, and the second dimension is for observations. It is two-dimensional even if there is only one observation.

Implementation details are left to concrete subtypes, and will depend on how the unary terms are parametrized. Note that indexing is performance-critical.

Concrete subtypes should implement getparameters, setparameters!, and showfields.

Examples

julia> M = ALsimple(Graph(4,4));
julia> typeof(M.unary)
FullUnary
julia> isa(M.unary, AbstractUnaryParameter)
true

source

Autologistic.CenteringKinds — Type.

CenteringKinds

An enumeration to facilitate choosing a form of centering for the model. Available choices are:

none: no centering (centering adjustment equals zero).

expectation: the centering adjustment is the expected value of the response under the assumption of independence (this is what has been used in the "centered autologistic model").

onehalf: a constant value of centering adjustment equal to 0.5.

source

Autologistic.FullPairwise — Type.

FullPairwise

A type representing an association matrix with a "saturated" parametrization–one parameter for each edge in the graph.

In this type, the association matrix for each observation is a symmetric matrix with the same pattern of nonzeros as the graph's adjacency matrix, but with arbitrary values in those locations. The package convention is to provide parameters as a vector of Float64. So getparameters and setparameters! use a vector of ne(G) values that correspond to the nonzero locations in the upper triangle of the adjacency matrix, in the same (lexicographic) order as edges(G).

The association matrix is stored as a SparseMatrixCSC{Float64,Int64} in the field Λ.

As with SimplePairwise, the association matrix can not be different for different observations. So while size returns a 3-dimensional result, the third index is ignored when accessing the array's elements.

source

Autologistic.SamplingMethods — Type.

SamplingMethods

An enumeration to facilitate choosing a method for sampling. Available choices are:

Gibbs TODO

perfect_bounding_chain TODO

perfect_reuse_samples TODO

perfect_reuse_seeds TODO

perfect_read_once TODO

source

Methods

Autologistic.fullPMF — Method.

fullPMF(M::AbstractAutologisticModel; indices=1:size(M.unary,2), force::Bool=false)

Compute the PMF of an AbstractAutologisticModel, and return a NamedTuple (:table, :partition).

For an AbstractAutologisticModel with $n$ variables and $m$ observations, :table is a $2^n×(n+1)×m$ array of Float64. Each page of the 3D array holds a probability table for an observation. Each row of the table holds a specific configuration of the responses, with the corresponding probability in the last column. In the $m=1$ case, :table is a 2D array.

Output :partition is a vector of normalizing constant (a.k.a. partition function) values. In the $m=1$ case, it is a scalar Float64.

Arguments

M: an autologistic model.

indices: indices of specific observations from which to obtain the output. By default, all observations are used.

force: calling the function with $n>20$ will throw an error unless force=true.

Examples

julia> M = ALRsimple(Graph(3,0),ones(3,1));
julia> pmf = fullPMF(M);
julia> pmf.table
8×4 Array{Float64,2}:
 -1.0  -1.0  -1.0  0.125
 -1.0  -1.0   1.0  0.125
 -1.0   1.0  -1.0  0.125
 -1.0   1.0   1.0  0.125
  1.0  -1.0  -1.0  0.125
  1.0  -1.0   1.0  0.125
  1.0   1.0  -1.0  0.125
  1.0   1.0   1.0  0.125
julia> pmf.partition
 8.0

source

Autologistic.sample — Function.

sample(
    M::AbstractAutologisticModel, 
    k::Int = 1;
    method::SamplingMethods = Gibbs,
    indices = 1:size(M.unary,2), 
    average::Bool = false, 
    config = nothing, 
    burnin::Int = 0,
    verbose::Bool = false
)

Draws k random samples from autologistic model M, and either returns the samples themselves, or the estimated probabilities of observing the "high" level at each vertex.

If the model has more than one observation, then k samples are drawn for each observation's parameter settings. To restrict the samples to a subset of observations, use argument indices.

For a model M with n vertices in its graph:

When average=false, the return value is n × length(indices) × k, with singleton dimensions dropped.

When average=true, the return value is n × length(indices), with singleton dimensions dropped.

Keyword Arguments

method is a member of the enum SamplingMethods, specifying which sampling method will be used. The default uses Gibbs sampling. Where feasible, it is recommended to use one of the perfect sampling alternatives. See SamplingMethods for more.

indices gives the indices of the observation to use for sampling. The default is all indices, in which case each sample is of the same size as M.responses.

average controls the form of the output. When average=true, the return value is the proportion of "high" samples at each vertex. (Note that this is not actually the arithmetic average of the samples, unless the coding is (0,1). Rather, it is an estimate of the probability of getting a "high" outcome.) When average=false, the full set of samples is returned.

config allows a starting configuration of the random variables to be provided. Only used if method=Gibbs. Any vector of the correct length, with two unique values, can be used as config. By default a random configuration is used.

burnin specifies the number of initial samples to discard from the results. Only used if method=Gibbs.

skip specifies how many samples to throw away between returned samples. Only used if method=Gibbs.

verbose controls output to the console. If true, intermediate information about sampling progress is printed to the console. Otherwise no output is shown.

Examples

julia> M = ALsimple(Graph(4,4));
julia> M.coding = (-2,3);
julia> r = sample(M,10);
julia> size(r)
(4, 10)
julia> sort(unique(r))
2-element Array{Float64,1}:
 -2.0
  3.0

source