API Reference

API Reference

Types and Constructors

Autologistic.AbstractAutologisticModel — Type.

AbstractAutologisticModel

Abstract type representing autologistic models. This type has methods defined for most operations one will want to perform, so that concrete subtypes should not have to define too many methods unless more specialized and efficient algorithms for the specific subtype.

All concrete subtypes should have the following fields:

responses::Array{Bool,2} – The binary observations. Rows are for nodes in the graph, and columns are for independent (vector) observations. It is a 2D array even if there is only one observation.

unary<:AbstractUnaryParameter – Specifies the unary part of the model.

pairwise<:AbstractPairwiseParameter – Specifies the pairwise part of the model (including the graph).

centering<:CenteringKinds – Specifies the form of centering used, if any.

coding::Tuple{T,T} where T<:Real – Gives the numeric coding of the responses.

labels::Tuple{String,String} – Provides names for the high and low states.

coordinates<:SpatialCoordinates – Provides 2D or 3D coordinates for each vertex in the graph (or nothing if no coordinates).

The following functions are defined for the abstract type, and are considered part of the type's interface (in this list, M of type inheriting from AbstractAutologisticModel).

getparameters(M) and setparameters!(M, newpars::Vector{Float64})

getunaryparameters(M) and setunaryparameters!(M, newpars::Vector{Float64})

getpairwiseparameters(M) and setpairwiseparameters!(M, newpars::Vector{Float64})

makecoded(M, Y)

centeringterms(M, kind::Union{Nothing,CenteringKinds})

pseudolikelihood(M)

negpotential(M)

fullPMF(M; indices, force::Bool)

marginalprobabilities(M; indices, force::Bool)

conditionalprobabilities(M; vertices, indices)

sample(M, k::Int, method::SamplingMethods, indices::Int, average::Bool, start, burnin::Int, verbose::Bool)

The sample() function is a wrapper for a variety of random sampling algorithms enumerated in SamplingMethods.

Examples

julia> M = ALsimple(Graph(4,4));
julia> typeof(M)
ALsimple{CenteringKinds,Int64,Nothing}
julia> isa(M, AbstractAutologisticModel)
true

source

Autologistic.SamplingMethods — Type.

SamplingMethods

An enumeration to facilitate choosing a method for sampling. Available choices are:

Gibbs TODO

perfect_bounding_chain TODO

perfect_reuse_samples TODO

perfect_reuse_seeds TODO

perfect_read_once TODO

source

Methods

Autologistic.fullPMF — Function.

fullPMF(M::AbstractAutologisticModel; indices=1:size(M.unary)[2], force::Bool=false)

Compute the PMF of an AbstractAutologisticModel, and return a NamedTuple (:table, :partition).

For an AutologisticModel with $n$ variables and $m$ observations, :table is a $2^n×(n+1)×m$ array of Float64. Each page of the 3D array holds a probability table for an observation. Each row of the table holds a specific configuration of the responses, with the corresponding probability in the last column. In the $m=1$ case, :table is a 2D array.

Output :partition is a vector of normalizing constant (a.k.a. partition function) values. In the $m=1$ case, it is a scalar Float64.

Arguments

M: an autologistic model.

indices: indices of specific observations from which to obtain the output. By default, all observations are used.

force: calling the function with $n>20$ will throw an error unless force=true.

Examples

julia> M = ALRsimple(Graph(3,0),ones(3,1));
julia> pmf = fullPMF(M);
julia> pmf.table
8×4 Array{Float64,2}:
 -1.0  -1.0  -1.0  0.125
 -1.0  -1.0   1.0  0.125
 -1.0   1.0  -1.0  0.125
 -1.0   1.0   1.0  0.125
  1.0  -1.0  -1.0  0.125
  1.0  -1.0   1.0  0.125
  1.0   1.0  -1.0  0.125
  1.0   1.0   1.0  0.125
julia> pmf.partition
 8.0

source

Autologistic.sample — Function.

sample(
    M::AbstractAutologisticModel, 
    k::Int = 1;
    method::SamplingMethods = Gibbs,
    index::Int = 1, 
    average::Bool = false, 
    start = nothing, 
    burnin::Int = 0,
    verbose::Bool = false
)

Draws k random samples from autologistic model M. For a model M with n vertices in its graph, the return value is an Array{Float64,1} of length n when average=true, and an n×k Array{Float64,2} when average=false.

Keyword Arguments

method is a member of the enum SamplingMethods, specifying which sampling method will be used. The default uses Gibbs sampling. Where feasible, it is recommended to use one of the perfect sampling alternatives. See SamplingMethods for more.

index gives the index of the observation to use for sampling.

average controls whether the return value is the proportion of "high" samples at each (when average=true) or the full set of samples (when average=false). Note that when the coding is not (0,1), the the return value is the estimated probability of getting a "high" outcome, not the arithmetic average of the samples.

start allows a starting configuration of the random variables to be provided. Only used if method=Gibbs. Any vector with two unique values can be used as start.

burnin specifies the number of initial samples to discard from the results. Only used if method=Gibbs.

verbose controls output to the console. If true, intermediate information about sampling progress is printed to the console. Otherwise no output is shown.

Examples

julia> M = ALsimple(Graph(4,4));
julia> M.coding = (-2,3);
julia> r = sample(M,10);
julia> size(r)
(4, 10)
julia> unique(r)
2-element Array{Float64,1}:
  3.0
 -2.0

source