API Reference

API Reference

Types and Constructors

TODO

Methods

Autologistic.fullPMF — Function.

fullPMF(M::ALmodel; replicates=nothing, force::Bool=false)

Compute the PMF of an ALmodel, and return a NamedTuple (:table, :partition).

For an ALmodel with $n$ observations and $m$ replicates, :table is a $2^n×(n+1)×m$ array of Float64. Each page of the 3D array holds a probability table for a replicate. Each row of the table holds a specific configuration of the responses, with the corresponding probability in the last column. In the $m=1$ case, :table is a 2D array.

Output :partition is a vector of normalizing constant (a.k.a. partition function) values. In the $m=1$ case, it is a scalar Float64.

Arguments

M::ALmodel: an autologistic model.

replicates=nothing: indices of specific replicates from which to obtain the output. By default, all replicates are used.

force::Bool=false: calling the function with $n>20$ will throw an error unless force=true.

Examples

julia> M = ALRsimple(Graph(3,0),ones(3,1));
julia> pmf = fullPMF(M);
julia> pmf.table
8×4 Array{Float64,2}:
 -1.0  -1.0  -1.0  0.125
 -1.0  -1.0   1.0  0.125
 -1.0   1.0  -1.0  0.125
 -1.0   1.0   1.0  0.125
  1.0  -1.0  -1.0  0.125
  1.0  -1.0   1.0  0.125
  1.0   1.0  -1.0  0.125
  1.0   1.0   1.0  0.125
julia> pmf.partition
 8.0

source