Reference

Reference

Index

Autologistic.ALfit

Autologistic.AbstractAutologisticModel

Autologistic.AbstractPairwiseParameter

Autologistic.CenteringKinds

Autologistic.FullPairwise

Autologistic.SamplingMethods

Autologistic.fullPMF

Autologistic.sample

Types and Constructors

Autologistic.ALfit — Type.

ALfit

A type to hold estimation output for autologistic models. Fitting functions return an object of this type.

Depending on the fitting method, some fields might not be set. Fields that are not used are set to nothing or to zero-dimensional arrays. The fields are:

estimate: A vector of parameter estimates

se: A vector of standard errors for the estimates

pvalues: A vector of p-values for testing the null hypothesis that the parameters equal zero (one-at-a time hypothesis tests).

CIs: A vector of 95% confidence intervals for the parameters (a vector of 2-tuples).

optim: the output of the call to optimize used to get the estimates.

Hinv (used by fit_ml!): The inverse of the Hessian matrix of the objective function, evaluated at the estimate.

nboot - (fit_pl!) number of bootstrap samples to use for error estimation

kwargs - (fit_pl!) string representation of extra keyword arguments passed in the call (a record of which arguments were passed to sample)

bootsamples - (fit_pl!) the bootstrap samples

bootestimates - (fit_pl!) the bootstrap parameter estimates

convergence - either a Boolean indicating optimization convergence ( for fit_ml!), or a vector of such values for the optimizations done to estimate bootstrap replicates.

The empty constructor ALfit() will initialize an object with all fields empty, so the needed fields can be filled afterwards.

source

Autologistic.AbstractAutologisticModel — Type.

AbstractAutologisticModel

Abstract type representing autologistic models. This type has methods defined for most operations one will want to perform, so that concrete subtypes should not have to define too many methods unless more specialized and efficient algorithms for the specific subtype.

All concrete subtypes should have the following fields:

responses::Array{Bool,2} – The binary observations. Rows are for nodes in the graph, and columns are for independent (vector) observations. It is a 2D array even if there is only one observation.

unary<:AbstractUnaryParameter – Specifies the unary part of the model.

pairwise<:AbstractPairwiseParameter – Specifies the pairwise part of the model (including the graph).

centering<:CenteringKinds – Specifies the form of centering used, if any.

coding::Tuple{T,T} where T<:Real – Gives the numeric coding of the responses.

labels::Tuple{String,String} – Provides names for the high and low states.

coordinates<:SpatialCoordinates – Provides 2D or 3D coordinates for each vertex in the graph.

The following functions are defined for the abstract type, and are considered part of the type's interface (in this list, M is of type inheriting from AbstractAutologisticModel).

getparameters(M) and setparameters!(M, newpars::Vector{Float64})

getunaryparameters(M) and setunaryparameters!(M, newpars::Vector{Float64})

getpairwiseparameters(M) and setpairwiseparameters!(M, newpars::Vector{Float64})

makecoded(M, Y)

centeringterms(M, kind::Union{Nothing,CenteringKinds})

pseudolikelihood(M)

negpotential(M)

fullPMF(M; indices, force::Bool)

marginalprobabilities(M; indices, force::Bool)

conditionalprobabilities(M; vertices, indices)

sample(M, k::Int, method::SamplingMethods, indices::Int, average::Bool, start, burnin::Int, verbose::Bool)

The sample() function is a wrapper for a variety of random sampling algorithms enumerated in SamplingMethods.

Examples

julia> M = ALsimple(Graph(4,4));
julia> typeof(M)
ALsimple{CenteringKinds,Int64,Nothing}
julia> isa(M, AbstractAutologisticModel)
true

source

Autologistic.AbstractPairwiseParameter — Type.

AbstractPairwiseParameter

Abstract type representing the pairwise part of an autologistic regression model.

All concrete subtypes should have the following fields:

G::SimpleGraph{Int} – The graph for the model.

count::Int – The number of observations.

In addition to getindex() and setindex!(), any concrete subtype P<:AbstractPairwiseParameter should also have the following methods defined:

getparameters(P), returning a Vector{Float64}

setparameters!(P, newpar::Vector{Float64}) for setting parameter values.

Note that indexing is performance-critical and should be implemented carefully in subtypes.

The intention is that each subtype should implement a different way of parameterizing the association matrix. The way parameters are stored and values computed is up to the subtypes.

This type inherits from AbstractArray{Float64, 3}. The third index is to allow for multiple observations. P[:,:,r] should return the association matrix of the rth observation in an appropriate subtype of AbstractMatrix. It is not intended that the third index will be used for range or vector indexing like P[:,:,1:5] (though this may work due to AbstractArray fallbacks).

Examples

julia> M = ALsimple(Graph(4,4));
julia> typeof(M.pairwise)
SimplePairwise
julia> isa(M.pairwise, AbstractPairwiseParameter)
true

source

Autologistic.CenteringKinds — Type.

CenteringKinds

An enumeration to facilitate choosing a form of centering for the model. Available choices are:

none: no centering (centering adjustment equals zero).

expectation: the centering adjustment is the expected value of the response under the assumption of independence (this is what has been used in the "centered autologistic model").

onehalf: a constant value of centering adjustment equal to 0.5.

source

Autologistic.FullPairwise — Type.

FullPairwise

A type representing an association matrix with a "saturated" parametrization–one parameter for each edge in the graph.

In this type, the association matrix for each observation is a symmetric matrix with the same pattern of nonzeros as the graph's adjacency matrix, but with arbitrary values in those locations. The package convention is to provide parameters as a vector of Float64. So getparameters and setparameters! use a vector of ne(G) values that correspond to the nonzero locations in the upper triangle of the adjacency matrix, in the same (lexicographic) order as edges(G).

The association matrix is stored as a SparseMatrixCSC{Float64,Int64} in the field Λ.

As with SimplePairwise, the association matrix can not be different for different observations. So while size returns a 3-dimensional result, the third index is ignored when accessing the array's elements.

source

Autologistic.SamplingMethods — Type.

SamplingMethods

An enumeration to facilitate choosing a method for sampling. Available choices are:

Gibbs TODO

perfect_bounding_chain TODO

perfect_reuse_samples TODO

perfect_reuse_seeds TODO

perfect_read_once TODO

source

Methods

Autologistic.fullPMF — Method.

fullPMF(M::AbstractAutologisticModel; indices=1:size(M.unary,2), force::Bool=false)

Compute the PMF of an AbstractAutologisticModel, and return a NamedTuple (:table, :partition).

For an AutologisticModel with $n$ variables and $m$ observations, :table is a $2^n×(n+1)×m$ array of Float64. Each page of the 3D array holds a probability table for an observation. Each row of the table holds a specific configuration of the responses, with the corresponding probability in the last column. In the $m=1$ case, :table is a 2D array.

Output :partition is a vector of normalizing constant (a.k.a. partition function) values. In the $m=1$ case, it is a scalar Float64.

Arguments

M: an autologistic model.

indices: indices of specific observations from which to obtain the output. By default, all observations are used.

force: calling the function with $n>20$ will throw an error unless force=true.

Examples

julia> M = ALRsimple(Graph(3,0),ones(3,1));
julia> pmf = fullPMF(M);
julia> pmf.table
8×4 Array{Float64,2}:
 -1.0  -1.0  -1.0  0.125
 -1.0  -1.0   1.0  0.125
 -1.0   1.0  -1.0  0.125
 -1.0   1.0   1.0  0.125
  1.0  -1.0  -1.0  0.125
  1.0  -1.0   1.0  0.125
  1.0   1.0  -1.0  0.125
  1.0   1.0   1.0  0.125
julia> pmf.partition
 8.0

source

Autologistic.sample — Function.

sample(
    M::AbstractAutologisticModel, 
    k::Int = 1;
    method::SamplingMethods = Gibbs,
    indices = 1:size(M.unary,2), 
    average::Bool = false, 
    start = nothing, 
    burnin::Int = 0,
    verbose::Bool = false
)

Draws k random samples from autologistic model M, and either returns the samples themselves, or the estimated probabilities of observing the "high" level at each vertex.

If the model has more than one observation, then k samples are drawn for each observation. To restrict the samples to a subset of observations, use argument indices.

For a model M with n vertices in its graph:

When average=false, the return value is n × length(indices) × k, with singleton dimensions dropped.

When average=true, the return value is n × length(indices), with singleton dimensions dropped.

Keyword Arguments

method is a member of the enum SamplingMethods, specifying which sampling method will be used. The default uses Gibbs sampling. Where feasible, it is recommended to use one of the perfect sampling alternatives. See SamplingMethods for more.

indices gives the indices of the observation to use for sampling. The default is all indices, in which case each sample is of the same size as M.responses.

average controls the form of the output. When average=true, the return value is the proportion of "high" samples at each vertex. (Note that this is not actually the arithmetic average of the samples, unless the coding is (0,1). Rather, it is an estimate of the probability of getting a "high" outcome.) When average=false, the full set of samples is returned.

start allows a starting configuration of the random variables to be provided. Only used if method=Gibbs. Any vector with two unique values can be used as start. By default a random configuration is used.

burnin specifies the number of initial samples to discard from the results. Only used if method=Gibbs.

skip specifies how many samples to throw away between returned samples. Only used if method=Gibbs.

verbose controls output to the console. If true, intermediate information about sampling progress is printed to the console. Otherwise no output is shown.

Examples

julia> M = ALsimple(Graph(4,4));
julia> M.coding = (-2,3);
julia> r = sample(M,10);
julia> size(r)
(4, 10)
julia> sort(unique(r))
2-element Array{Float64,1}:
 -2.0
  3.0

source